Update dependabot frequency and remove Jenkins
Dependabot should skip the path version updates for Spring Boot.
Prior to this change the patch version on all dependencies was excluded.
This change specifies org.springframework.* specifically in the
Gradle folders in order to allow the dependency
io.spring.dependency-management to be updated on the patch release.
There is no maven equivalent for this, so only the Gradle
exclusion was changed.

This change also removes the test folder which was applicable only
to the Jenkins CI environment, which is being decomissioned.